Update DietOjalgo.groovy
When you call minimise or maximise, everything with a weight is aggregated (weighted) together, and the value of that combined expression is minimised/maximised.

That means the same objective can be declared in different ways. For an LP model the most direct way is to simply set weights on the variables.

Due to how ojAlgo works internally, it is more efficient to declare as much as possible on the variables directly, rather than to create separate expressions for everything.